Add files via upload
We use a logistic regression model to predict the willingness of HIV people and healthcare users to accept telemedicine in public hospitals in Buenos Aires. The aim was to develop a model that can identify those who are willing to use this strategy so administrative assistants do not have to offer this modality to all patients. For that reason, we developed a logistic regression model with thereshold. This model is not proper to identify those who are not willing to use telemedicine. We compare this model with ensemble models (XG boost, random forest and Gradient Boost) and logistic regression better solved our business problem.